+++ packages/browser-wallet-api-helpers/src/wallet-api-types.ts
@@ -76,6 +76,11 @@ interface MainWalletApi {
      */
     connect(): Promise<string | undefined>;
 
+    /**
+     * Returns some connected account, prioritizing the most recently selected. Resolves with account address or undefined if there are no connected account.
+     */
+    getMostRecentlySelectedAccount(): Promise<string | undefined>;
+
     removeAllListeners(event?: EventType | string | undefined): this;
 
     getJsonRpcClient(): JsonRpcClient;